fix(container): allow setSize to work with widths and heights
Currently the `setSize` method on containers only works with heights but
accepts 2 parameter `width, height`. If the container is part of a row
the width is used to resize the widths of all content items. If the
container is part of a column the height is used to resize the height of
all content items.