Fix abusive calls to AbstractContentItem._$destroy()
There seems to be a flaw around the `_$destroy()` and `callDownwards()` methods. When destroying a layout with 20+ components, it takes an abnormably long time. When profiling, I see that the `_$destroy()` method is called repeatedly for the same item.

To resolve the problem , the method `_$destroy()` method is split in two, one public method that destroys children recursively and one private method that does not propagate downwards.