Use shallowRef in the Vue example of README.md
If an object is assigned as a Vue `ref`'s value, the object is made deeply reactive by the `reactive` method.
This might introduce some issues in some circumstances, because a reactive wrapper has been created.
To avoid potential issues, `ref` can be replaced with `shallowRef` in the readme example. `shallowRef` will track the value change, but it won't make its value reactive.