Fix implementation of stripTags.  Don't use it for Tab.setTitle, though
If the input to stripTags is allowed to be HTML, then the old
implementation failed to handle character entities.  Hence no
way to create a title containing '<' or '>'.  A simple fix is
to parse parse the input as HTML (using innerHTML on a temporary element)
and extract the text (using innerText).

Also, for Tab.setTitle, it is useful to actually allow HTML,
so don't use stripTags.