Disable user-select of drag handles.
I was experiencing a problem when using my mouse to resize a splitter, some text in the content box was getting selected.  It happens in the short moment before the mouse has moved far enough to trigger the resizing action.  Simply disabling user-select on it fixes it, and appears to fit alongside touch-action none, which was already existent.

Tested on Firefox and Chromium on Linux.  Problem did not appear on Chromium.  The fix functions on both.