Change _allowableTargets logic in drag-listener
With long-title there may be nested elements in an lm_title,
which broke the logic for checking allowable drag targets.
Fix this by scapping the _allowableTarget variable and instead setting
a 'draggable' attribute on elements, as in the drag-and-drop api.
(We don't yet use the standad drag-and-drop api, but hope to
in the future, so we might as well use the same convention.)