Catch iframe pointer-events in other top-level windows
When a drag starts we disable pointer-events in iframes so we can
we can follow mouse and drag events.  The complication is
when the drag starts in one top-level window, but drag-over freedback
and possible drop cn happen on other top-level windows. This change
provides a event-hook that applications can use to notify other windows.

Also some other minor tweaks.